+++ SplatTagCore/DivType.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace SplatTagCore
+{
+  [Serializable]
+  public enum DivType
+  {
+    Unknown,
+    LUTI,
+    EBTV
+  }
+}