+++ SplatTagCore/Player.cs
@@ -23,10 +23,20 @@ namespace SplatTagCore
     private List<string> names = new List<string>();
 
     /// <summary>
+    /// Back-store for the sources of this player.
+    /// </summary>
+    private List<string> sources = new List<string>();
+
+    /// <summary>
     /// Back-store for the team ids for this player. The first element is the current team.
     /// A 0 represents no team.
     /// </summary>
     private List<long> teams = new List<long>();
+
+    /// <summary>
+    /// Back-store for the weapons that the player uses (if any).
+    /// </summary>
+    private List<string> weapons = new List<string>();
 
     [JsonProperty("Names", Required = Required.Always)]
     /// <summary>
@@ -48,6 +58,7 @@ namespace SplatTagCore
       }
     }
 
+    [JsonIgnore]
     /// <summary>
     /// The last known used name for the player
     /// </summary>
@@ -74,6 +85,17 @@ namespace SplatTagCore
       set => teams = new List<long>(value ?? new long[0]);
     }
 
+    [JsonIgnore]
+    /// <summary>
+    /// The old teams this player has played for.
+    /// A 0 represents no team.
+    /// </summary>
+    public IEnumerable<long> OldTeams
+    {
+      get => teams?.Skip(1).ToArray();
+    }
+
+    [JsonIgnore]
     /// <summary>
     /// The current team id this player plays for, or 0 if not set.
     /// </summary>
@@ -90,11 +112,45 @@ namespace SplatTagCore
       }
     }
 
+    [JsonProperty("Weapons", Required = Required.Default)]
+    /// <summary>
+    /// The weapons this player uses.
+    /// </summary>
+    public string[] Weapons
+    {
+      get => weapons.ToArray();
+      set
+      {
+        weapons = new List<string>();
+        foreach (string s in value)
+        {
+          if (!string.IsNullOrWhiteSpace(s) && !weapons.Contains(s, StringComparison.OrdinalIgnoreCase))
+          {
+            weapons.Add(s);
+          }
+        }
+      }
+    }
+
     [JsonProperty("Sources", Required = Required.Default)]
     /// <summary>
     /// Get or Set the current sources that make up this Player instance.
     /// </summary>
-    public List<string> Sources { get; set; } = new List<string>();
+    public string[] Sources
+    {
+      get => sources.ToArray();
+      set
+      {
+        sources = new List<string>();
+        foreach (string s in value)
+        {
+          if (!string.IsNullOrWhiteSpace(s) && !names.Contains(s))
+          {
+            sources.Add(s);
+          }
+        }
+      }
+    }
 
     [JsonProperty("Id", Required = Required.Always)]
     /// <summary>
@@ -102,30 +158,62 @@ namespace SplatTagCore
     /// </summary>
     public uint Id { get; set; }
 
+    [JsonProperty("SendouId", Required = Required.Default)]
+    /// <summary>
+    /// The Sendou database Id of the player.
+    /// Null by default.
+    /// </summary>
+    public ulong? SendouId { get; set; }
+
+    [JsonProperty("DiscordId", Required = Required.Default)]
+    /// <summary>
+    /// The Discord database Id of the player.
+    /// Null by default.
+    /// </summary>
+    public ulong? DiscordId { get; set; }
+
     [JsonProperty("DiscordName", Required = Required.Default)]
     /// <summary>
     /// Get or Set the Discord Name.
     /// Null by default.
     /// </summary>
-    public string DiscordName { get; set; } = null;
+    public string DiscordName { get; set; }
 
     [JsonProperty("FriendCode", Required = Required.Default)]
     /// <summary>
     /// Get or Set the Friend Code.
     /// Null by default.
     /// </summary>
-    public string FriendCode { get; set; } = null;
+    public string FriendCode { get; set; }
+
+    [JsonProperty("Country", Required = Required.Default)]
+    /// <summary>
+    /// Get or Set the Country.
+    /// Null by default.
+    /// </summary>
+    public string Country { get; set; }
+
+    [JsonProperty("Top500", Required = Required.Default)]
+    /// <summary>
+    /// Get or Set Top 500 flag.
+    /// False by default.
+    /// </summary>
+    public bool Top500 { get; set; }
 
     /// <summary>
     /// Merge this player with another (newer) player instance
     /// </summary>
     /// <param name="otherPlayer"></param>
+    /// <exception cref="ArgumentNullException"><paramref name="otherPlayer"/> is <c>null</c>.</exception>
     public void Merge(Player otherPlayer)
     {
+      if (otherPlayer == null) throw new ArgumentNullException(nameof(otherPlayer));
+      if (ReferenceEquals(this, otherPlayer)) return;
+
       // Merge the players.
       // Iterates the other stack in reverse order so older teams are pushed first
       // so the most recent end up first in the stack.
-      var reverseTeams = otherPlayer.teams;
+      var reverseTeams = otherPlayer.teams.ToList();
       reverseTeams.Reverse();
       foreach (uint t in reverseTeams)
       {
@@ -146,9 +234,9 @@ namespace SplatTagCore
       // Merge the player's name(s).
       // Iterates the other stack in reverse order so older names are pushed first
       // so the most recent end up first in the stack.
-      var reversePlayers = otherPlayer.names;
+      var reversePlayers = otherPlayer.names.ToList();
       reversePlayers.Reverse();
-      foreach (string n in reversePlayers)
+      foreach (string n in reversePlayers.Where(s => !string.IsNullOrWhiteSpace(s)))
       {
         string foundName = this.names.Find(playerNames => playerNames.Equals(n, StringComparison.OrdinalIgnoreCase));
 
@@ -164,25 +252,56 @@ namespace SplatTagCore
       }
 
       // Merge the sources.
-      foreach (string source in otherPlayer.Sources)
+      foreach (string source in otherPlayer.Sources.Where(s => !string.IsNullOrWhiteSpace(s)))
       {
-        string foundSource = Sources.Find(sources => sources.Equals(source, StringComparison.OrdinalIgnoreCase));
+        string foundSource = this.sources.Find(sources => sources.Equals(source, StringComparison.OrdinalIgnoreCase));
 
         if (foundSource == null)
         {
-          Sources.Add(source);
+          sources.Add(source);
+        }
+      }
+
+      // Merge the weapons.
+      foreach (string weapon in otherPlayer.Weapons.Where(s => !string.IsNullOrWhiteSpace(s)))
+      {
+        string foundWeapon = weapons.Find(wep => weapon.Equals(wep, StringComparison.OrdinalIgnoreCase));
+
+        if (foundWeapon == null)
+        {
+          weapons.Add(weapon);
         }
       }
 
       // Merge the misc data
-      if (otherPlayer.FriendCode != null)
+      if (!string.IsNullOrWhiteSpace(otherPlayer.FriendCode))
       {
         this.FriendCode = otherPlayer.FriendCode;
       }
 
-      if (otherPlayer.DiscordName != null)
+      if (!string.IsNullOrWhiteSpace(otherPlayer.DiscordName))
       {
         this.DiscordName = otherPlayer.DiscordName;
+      }
+
+      if (otherPlayer.DiscordId != null)
+      {
+        this.DiscordId = otherPlayer.DiscordId;
+      }
+
+      if (!string.IsNullOrWhiteSpace(otherPlayer.Country))
+      {
+        this.Country = otherPlayer.Country;
+      }
+
+      if (otherPlayer.SendouId != null)
+      {
+        this.SendouId = otherPlayer.SendouId;
+      }
+
+      if (otherPlayer.Top500)
+      {
+        this.Top500 = true;
       }
     }