+++ SplatTagDatabase/Merger.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -9,11 +10,33 @@ namespace SplatTagDatabase
 {
   internal static class Merger
   {
+    public static void FinalisePlayers(IDictionary<uint, Player> playersToMutate)
+    {
+      if (playersToMutate == null) return;
+
+      var keys = playersToMutate.Keys.Reverse().ToArray();
+      foreach (var playerId in keys)
+      {
+        var playersMinusCurrent = playersToMutate.Values.Where(p => p.Id != playerId);
+        Debug.Assert(playersMinusCurrent.Count() == playersToMutate.Count - 1);
+        var testPlayer = playersToMutate[playerId];
+        Player foundPlayer = FindSamePlayer(playersMinusCurrent, testPlayer, null);
+        if (foundPlayer != null)
+        {
+          Debug.Assert(foundPlayer != testPlayer, $"Player to merge is the same as the one being merged: {foundPlayer.Name} (Id {foundPlayer.Id}) equiv to {testPlayer.Name} (Id [{playerId}] == {testPlayer.Id})");
+          foundPlayer.Merge(testPlayer);
+          playersToMutate.Remove(playerId);
+        }
+      }
+    }
+
     /// <summary>
     /// Merge the loaded players into the current players list.
     /// </summary>
     public static void MergePlayers(IDictionary<uint, Player> playersToMutate, IEnumerable<Player> incomingPlayers, IReadOnlyDictionary<long, Team> incomingTeams)
     {
+      if (playersToMutate == null || incomingPlayers == null) return;
+
       // Add if the player is new (by name) and assign them with a new id
       // Otherwise, match the found team with its id, based on name.
       ConcurrentBag<Player> playersToAdd = new ConcurrentBag<Player>();
@@ -27,19 +50,29 @@ namespace SplatTagDatabase
 
       Parallel.ForEach(incomingPlayers, (importPlayer) =>
       {
-        // Correct the ids
-        importPlayer.Teams = importPlayer.Teams.Select(idToMerge => incomingTeams[idToMerge].Id);
-
-        Player foundPlayer = FindSamePlayer(playersToMutate.Values, importPlayer, transformedPlayerNames);
-
-        if (foundPlayer == null)
+        try
         {
-          playersToAdd.Add(importPlayer);
+          // Correct the ids
+          if (incomingTeams != null)
+          {
+            importPlayer.Teams = importPlayer.Teams.Select(idToMerge => incomingTeams[idToMerge].Id);
+          }
+
+          Player foundPlayer = FindSamePlayer(playersToMutate.Values, importPlayer, transformedPlayerNames);
+
+          if (foundPlayer == null)
+          {
+            playersToAdd.Add(importPlayer);
+          }
+          else
+          {
+            importPlayer.Id = foundPlayer.Id;
+            playersToMerge.Add((foundPlayer, importPlayer));
+          }
         }
-        else
+        catch (Exception ex)
         {
-          importPlayer.Id = foundPlayer.Id;
-          playersToMerge.Add((foundPlayer, importPlayer));
+          Console.Error.WriteLine($"Error in importing player {importPlayer}: {ex}");
         }
       });
 
@@ -62,16 +95,24 @@ namespace SplatTagDatabase
     /// <param name="playersToMutate">The players to search</param>
     /// <param name="testPlayer">The player instance to try and find</param>
     /// <returns>The matched player, or null if new</returns>
-    private static Player FindSamePlayer(ICollection<Player> playersToMutate, Player testPlayer, Dictionary<Player, string[]> transformedPlayerNames)
+    private static Player FindSamePlayer(IEnumerable<Player> playersToMutate, Player testPlayer, IEnumerable<KeyValuePair<Player, string[]>> transformedPlayerNames)
     {
-      bool tryMatchFcs = testPlayer.FriendCode != null;
-      bool tryMatchDiscord = testPlayer.DiscordName != null;
+      bool tryMatchFcs = !string.IsNullOrEmpty(testPlayer.FriendCode);
+      bool tryMatchDiscord = !string.IsNullOrEmpty(testPlayer.DiscordName);
+      bool tryMatchDiscordId = testPlayer.DiscordId != null;
 
       // Quick test all FCs and Discord names.
       Player foundSame = playersToMutate.FirstOrDefault(p =>
       {
         // Test if the Switch FC's match.
         if (tryMatchFcs && p.FriendCode?.Equals(testPlayer.FriendCode) == true)
+        {
+          // They do.
+          return true;
+        }
+
+        // Test if the Discord Ids match.
+        if (tryMatchDiscordId && p.DiscordId?.Equals(testPlayer.DiscordId) == true)
         {
           // They do.
           return true;
@@ -91,19 +132,40 @@ namespace SplatTagDatabase
       // Long test for names with transformation.
       HashSet<string> testPlayerTransformedNames = new HashSet<string>(testPlayer.Names.Select(n => n.Replace(" ", "").TransformString().ToLowerInvariant()));
 
-      return transformedPlayerNames.FirstOrDefault(tPair =>
+      // Use the cached version if populated, else work out this player.
+      if (transformedPlayerNames == null)
       {
-        // Test if the name matches the names that we know this player by.
-        foreach (string knownName in tPair.Value)
+        return playersToMutate.FirstOrDefault(p =>
         {
-          if (testPlayerTransformedNames.Contains(knownName))
+          // Replace spaces because people adding tags or starting with space messes up same-name detection.
+          // Test if the name matches the names that we know this player by.
+          foreach (string knownName in p.Names.Select(n => n.Replace(" ", "").TransformString().ToLowerInvariant()))
           {
-            return true;
+            if ((!string.IsNullOrWhiteSpace(knownName)) && (testPlayerTransformedNames.Contains(knownName)))
+            {
+              return true;
+            }
           }
-        }
 
-        return false;
-      }).Key;
+          return false;
+        });
+      }
+      else
+      {
+        return transformedPlayerNames.FirstOrDefault(tPair =>
+        {
+          // Test if the name matches the names that we know this player by.
+          foreach (string knownName in tPair.Value)
+          {
+            if ((!string.IsNullOrWhiteSpace(knownName)) && (testPlayerTransformedNames.Contains(knownName)))
+            {
+              return true;
+            }
+          }
+
+          return false;
+        }).Key;
+      }
     }
 
     /// <summary>