+++ SplatTagDatabase/MultiDatabase.cs
@@ -23,18 +23,25 @@ namespace SplatTagDatabase
       Dictionary<long, Team> teams = new Dictionary<long, Team>();
       foreach (var db in importers)
       {
+        Dictionary<long, Team> teamDictionaryPreMerge = new Dictionary<long, Team>();
         var (loadedPlayers, loadedTeams) = db.Load();
-        var teamDictionaryPreMerge = loadedTeams.ToDictionary(team => team.Id, team => team);
-        Merger.MergeTeams(teams, loadedTeams);
+        if (loadedTeams != null)
+        {
+          teamDictionaryPreMerge = loadedTeams.ToDictionary(team => team.Id, team => team);
+          Merger.MergeTeams(teams, loadedTeams);
+        }
         Merger.MergePlayers(players, loadedPlayers, teamDictionaryPreMerge);
       }
+
+      // Perform a final merge.
+      Merger.FinalisePlayers(players);
       return (players.Values.ToArray(), teams.Values.ToArray());
     }
 
     public void Save(IEnumerable<Player> players, IEnumerable<Team> teams)
     {
       // Offload to a Json database instead.
-      SplatTagJsonDatabase jsonDb = new SplatTagJsonDatabase(saveDirectory);
+      SplatTagJsonSnapshotDatabase jsonDb = new SplatTagJsonSnapshotDatabase(saveDirectory);
       jsonDb.Save(players, teams);
       importers.Clear();
       importers.Add(jsonDb);