+++ SplatTagUI/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using SplatTagCore;
+﻿using Newtonsoft.Json;
+using SplatTagCore;
 using SplatTagDatabase;
 using System;
 using System.Collections.Generic;
@@ -24,41 +25,64 @@ namespace SplatTagUI
     private const int MIN_CHARACTERS_FOR_TIMER_SKIP = 4;
 
     internal static readonly SplatTagController splatTagController;
-    private static readonly SplatTagJsonDatabase jsonDatabase;
-    private static readonly MultiDatabase splatTagDatabase;
-    private static readonly GenericFilesImporter multiSourceImporter;
     private static readonly string saveFolder = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "SplatTag");
+    private static readonly GenericFilesImporter sourcesImporter;
     private readonly Timer smoothSearchDelayTimer;
     private readonly SynchronizationContext context;
 
     /// <summary>
     /// Version string to display.
     /// </summary>
-    public string Version => "Version 0.0.13";
+    public string Version => "Version 0.0.15";
 
     /// <summary>
     /// Version tooltip string to display.
     /// </summary>
-    public string VersionToolTip => "v0.0.13: Better div support. Better sameness detection.\nv0.0.12: Added command-line. \nv0.0.11: Added Twitter handle capability. \nv0.0.10: Supports InkTV. Pulls friend codes from names. Corrected bugs around incomplete tournament data. Added a browse button to the file loader. \nv0.0.09: Now searches FCs and Discord tags. Teams now have a highest div'd label. Stability improvements.";
+    public string VersionToolTip =>
+      "v0.0.15: Sendou details \n" +
+      "v0.0.14: Now works off of snapshots to drastically improve load time. \n" +
+      "v0.0.13: Better div support. Better sameness detection.\n" +
+      "v0.0.12: Added command-line. \n" +
+      "v0.0.11: Added Twitter handle capability. \n" +
+      "v0.0.10: Supports InkTV. Pulls friend codes from names. Corrected bugs around incomplete tournament data. Added a browse button to the file loader. \n" +
+      "v0.0.09: Now searches FCs and Discord tags. Teams now have a highest div'd label. Stability improvements.";
 
     static MainWindow()
     {
-      jsonDatabase = new SplatTagJsonDatabase(saveFolder);
-      multiSourceImporter = new GenericFilesImporter(saveFolder);
-      splatTagDatabase = new MultiDatabase(saveFolder, jsonDatabase, multiSourceImporter);
-      splatTagController = new SplatTagController(splatTagDatabase);
+      SplatTagJsonSnapshotDatabase snapshotDatabase = new SplatTagJsonSnapshotDatabase(saveFolder);
+      splatTagController = new SplatTagController(snapshotDatabase);
       splatTagController.Initialise();
+
+      // If we were able to load from a snapshot then we don't need the other importers.
+      // Otherwise, do the processing and record the snapshot.
+      if (splatTagController.MatchPlayer(null).Length == 0)
+      {
+        sourcesImporter = new GenericFilesImporter(saveFolder);
+        MultiDatabase splatTagDatabase = new MultiDatabase(saveFolder, sourcesImporter);
+        splatTagController = new SplatTagController(splatTagDatabase);
+        splatTagController.Initialise();
+
+        // Now that we've initialised, take a snapshot of everything.
+        snapshotDatabase.Save(splatTagController.MatchPlayer(null), splatTagController.MatchTeam(null));
+      }
     }
 
     public MainWindow()
     {
       InitializeComponent();
+      Title = $"Slapp - {splatTagController.MatchPlayer(null).Length} Players and {splatTagController.MatchTeam(null).Length} Teams loaded!";
 
       // Initialise the delay timer if we have a UI context, otherwise don't use the timer.
       context = SynchronizationContext.Current;
       if (context != null)
       {
         smoothSearchDelayTimer = new Timer(TimerExpired);
+      }
+
+      // If we've loaded from a snapshot, then hide the setup button(s).
+      if (sourcesImporter == null)
+      {
+        otherFunctionsGrid.Visibility = Visibility.Collapsed;
       }
 
       // Now we've initialised, hook up the check changed.
@@ -138,12 +162,14 @@ namespace SplatTagUI
 
     private void DataBaseButton_Click(object sender, RoutedEventArgs e)
     {
-      DatabaseEditor databaseEditor = new DatabaseEditor(splatTagController, multiSourceImporter.Sources);
+      if (sourcesImporter == null) return;
+
+      DatabaseEditor databaseEditor = new DatabaseEditor(splatTagController, sourcesImporter.Sources);
       bool? result = databaseEditor.ShowDialog();
       if (result == true)
       {
         // Editor accepted, save the sources list.
-        multiSourceImporter.SaveSources(databaseEditor.DatabaseSources);
+        sourcesImporter.SaveSources(databaseEditor.DatabaseSources);
 
         // And load the new database.
         splatTagController.LoadDatabase();
@@ -204,21 +230,24 @@ namespace SplatTagUI
   {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-      IEnumerable<Team> teams;
+      if (value == null) return "(not set)";
+
+      IEnumerable<Team> oldTeams;
       if (value is Player p)
       {
-        teams = p.Teams.Select(id => MainWindow.splatTagController.GetTeamById(id));
+        oldTeams = p.OldTeams.Select(id => MainWindow.splatTagController?.GetTeamById(id) ?? Team.NoTeam);
       }
       else if (value is IEnumerable<Team> t)
       {
-        teams = t;
+        oldTeams = t;
       }
       else
       {
         throw new InvalidDataException("Unknown type to convert: " + value.GetType());
       }
 
-      return teams.GetOldTeamsStrings();
+      string separator = parameter?.ToString() ?? ", ";
+      return string.Join(separator, oldTeams);
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => throw new NotImplementedException();
@@ -241,6 +270,29 @@ namespace SplatTagUI
       }
 
       return t.GetBestTeamPlayerDivString(MainWindow.splatTagController);
+    }
+
+    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => throw new NotImplementedException();
+  }
+
+  public class Top500ToString : IValueConverter
+  {
+    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+      bool top500;
+      if (value is Player p)
+      {
+        top500 = p.Top500;
+      }
+      else if (value is bool)
+      {
+        top500 = (bool)value;
+      }
+      else
+      {
+        throw new InvalidDataException("Unknown type to convert: " + value.GetType());
+      }
+      return top500 ? "👑" : "";
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => throw new NotImplementedException();
@@ -271,13 +323,19 @@ namespace SplatTagUI
   {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
+      string separator = ", ";
+      if (parameter != null)
+      {
+        separator = parameter.ToString();
+      }
+
       if (value is string[] values1)
       {
-        return string.Join(", ", values1);
+        return string.Join(separator, values1);
       }
       else if (value is IEnumerable<string> values2)
       {
-        return string.Join(", ", values2);
+        return string.Join(separator, values2);
       }
       else if (value is string values3)
       {
@@ -296,14 +354,10 @@ namespace SplatTagUI
   {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-      bool isValid;
-      if (value is string str)
+      bool isValid = !string.IsNullOrWhiteSpace(value?.ToString());
+      if (value is IEnumerable<string> collection)
       {
-        isValid = !string.IsNullOrWhiteSpace(str);
-      }
-      else
-      {
-        isValid = false;
+        isValid = collection.Any(s => !string.IsNullOrWhiteSpace(s?.ToString()));
       }
 
       return new BooleanToVisibilityConverter().Convert(isValid, targetType, parameter, culture);